refactor: moves date input to common components
