+++ src/modules/search-time/selector/selector.module.css
@@ -83,124 +83,6 @@
   margin-top: 0.375rem;
 }
 
-.dateSelector,
-.timeSelector {
-  --height: 2.75rem;
-  border-radius: var(--border-radius-regular);
-  overflow: hidden;
-  background-color: var(--static-background-background_0-background);
-  display: flex;
-  justify-content: space-between;
-}
-
-.dateSelector label,
-.timeSelector label,
-.timeSelectorLabel {
-  padding: var(--spacings-medium);
-  padding-right: 0;
-  height: var(--height);
-  min-width: 3rem;
-}
-
-.dateSelector input[type='date'],
-.timeSelector input[type='time'] {
-  height: var(--height);
-  padding: var(--spacings-medium);
-  border: none;
-  background-color: transparent;
-  color: var(--static-background-background_0-text);
-  flex: 1;
-  border-top-right-radius: 0.75rem;
-  border-bottom-right-radius: 0.75rem;
-}
-
-.calendarSelectorGroup {
-  display: flex;
-}
-.timeSelectorInput {
-  height: var(--height);
-  padding: var(--spacings-medium);
-  color: var(--static-background-background_0-text);
-  border-top-right-radius: 0.75rem;
-  border-bottom-right-radius: 0.75rem;
-  display: flex;
-}
-.timeSelectorSegment:focus {
-  background: var(--static-background-background_2-background);
-}
-
-.calendarButton {
-  background: var(--static-background-background_0-background);
-  border: none;
-  border-top-right-radius: 0.75rem;
-  border-bottom-right-radius: 0.75rem;
-  padding: var(--spacings-xSmall) var(--spacings-medium);
-  cursor: pointer;
-  margin: 2px 0;
-}
-
-.dateSelector input[type='date']:focus,
-.timeSelector input[type='time']:focus {
-  outline: 0;
-}
-
-.dateSelector:focus-within,
-.timeSelector:focus-within {
-  box-shadow: inset 0 0 0 var(--border-width-medium)
-    var(--interactive-interactive_2-outline-background);
-}
-
-.calendarDialog {
-  padding: var(--spacings-xSmall);
-  background: var(--static-background-background_0-background);
-  box-shadow: 0.2rem 0.2rem 0.3rem rgba(0, 0, 0, 0.15);
-}
-.calendarDialog__header {
-  display: flex;
-  gap: var(--spacings-medium);
-  justify-content: center;
-  align-items: center;
-  padding: var(--spacings-small);
-}
-.calendarDialog__title {
-  composes: typo-heading__component from '@atb/theme/typography.module.css';
-}
-.calendarDialog__headerButtons {
-  background: none;
-  border: none;
-  padding: var(--spacings-xSmall) var(--spacings-medium);
-  cursor: pointer;
-}
-
-.calendarGrid {
-  width: 100%;
-}
-
-.calendarGrid__cell {
-  padding: var(--spacings-xSmall) var(--spacings-medium);
-  cursor: pointer;
-  text-align: center;
-}
-.calendarGrid__cell[data-disabled] {
-  opacity: 0.3;
-  cursor: default;
-}
-.calendarGrid__cell[data-selected] {
-  background-color: var(--interactive-interactive_2-active-background);
-  color: var(--interactive-interactive_2-active-text);
-  font-weight: bold;
-  border-radius: var(--border-radius-small);
-}
-
-:global(.dark)
-  .dateSelector
-  input[type='date']::-webkit-calendar-picker-indicator,
-:global(.dark)
-  .timeSelector
-  input[type='time']::-webkit-calendar-picker-indicator {
-  filter: invert(1);
-}
-
 @media (max-width: 650px) {
   .dateAndTimeSelectors {
     flex-direction: column;
